Modifications list is now a lazy property, and it is instantiated on the Commit instead of in RepositoryMining
